+++ org.picfight.exchbot.lambda/src/org/picfight/exchbot/lambda/backend/ExchangeBackEndArgs.java
@@ -4,7 +4,7 @@ package org.picfight.exchbot.lambda.backend;
 public class ExchangeBackEndArgs {
 
 	public String host;
-	public int port;
+	public long port;
 	public String access_key;
 
 }
